Fix for hash function and split binding.gyp

+++ hashFn.h
@@ -12,8 +12,8 @@
 // Implements a Rabin fingerprint hash function
 class HashFn {
  public:
-  // Initialize a HashFn with the prime p which is used as the base of the
-  // Rabin fingerprint algorithm
+  // Initialize a HashFn with the prime p which is used as the base of the Rabin
+  // fingerprint algorithm
   explicit HashFn(int p) {
     this->p = p;
   }
@@ -22,13 +22,15 @@ class HashFn {
       unsigned char lastCharCode, uint64_t lastHash) {
     // See the abracadabra example:
     // https://en.wikipedia.org/wiki/Rabin%E2%80%93Karp_algorithm
-    return (lastHash - lastCharCode * pow(p, len - 1)) * p + input[len - 1];
+    return (lastHash - lastCharCode * static_cast<uint64_t>(
+          pow(static_cast<long double>(p), len - 1))) * p + input[len - 1];
   }
 
   virtual uint64_t operator()(const char *input, int len) {
     uint64_t total = 0;
     for (int i = 0; i < len; i++) {
-      total += input[i] * pow(p, len - i - 1);
+      total += input[i] * static_cast<uint64_t>(
+          pow(static_cast<long double>(p), len - i - 1));
     }
     return total;
   }
